fix(focus): silence build warnings in controls, logging, and views
- Add the missing Event timestamp initializer to VALUE_CHANGED events in
  Checkbox, RadioButton, and Slider
- Use PRId32 for the tag field in View::description and include
  <inttypes.h>
- Cast the bidi-class ternary to uint8_t in CanvasView::renderBidiLine
- FOCUS_LOGI/D/V stubs now reference their arguments via an if(0)
  fprintf so variables used only by compiled-out log levels are not
  flagged unused, and format strings stay type-checked

+++ src/views/CanvasView.cpp
@@ -607,7 +607,7 @@ void CanvasView::renderBidiLine(UNICODE_CODEPOINT *codepoints, size_t lineStart,
     for (size_t i = 0; i < lineLen; i++) {
         uint8_t bc = resolved[i];
         if (bc == BIDI_NSM) {
-            resolved[i] = (i > 0) ? resolved[i - 1] : (paragraphDir == -1 ? BIDI_R : BIDI_L);
+            resolved[i] = (i > 0) ? resolved[i - 1] : (uint8_t)(paragraphDir == -1 ? BIDI_R : BIDI_L);
         }
     }
     for (size_t i = 0; i < lineLen; i++) {
